separate the home screen and the landing screen
